embed dtos file
